feat: Add v3 version of the node
