+++ packages/nodes-base/nodes/Cal/test/CalTrigger.node.test.ts
@@ -4,6 +4,7 @@ import type { IHookFunctions, IWebhookFunctions } from 'n8n-workflow';
 import { CalTrigger } from '../CalTrigger.node';
 import { verifySignature } from '../CalTriggerHelpers';
 import { calApiRequest } from '../GenericFunctions';
+import type { CalTriggerV1 } from '../v1/CalTriggerV1.node';
 import type { Mock, Mocked } from 'vitest';
 import type * as _importType0 from 'crypto';
 
@@ -15,7 +16,7 @@ vi.mock('crypto', async () => ({
 }));
 
 describe('CalTrigger', () => {
-	let trigger: CalTrigger;
+	let trigger: CalTriggerV1;
 	let mockHookFunctions: Pick<
 		Mocked<IHookFunctions>,
 		'getNodeWebhookUrl' | 'getNodeParameter' | 'getWorkflowStaticData' | 'helpers'
@@ -27,7 +28,7 @@ describe('CalTrigger', () => {
 
 	beforeEach(() => {
 		vi.clearAllMocks();
-		trigger = new CalTrigger();
+		trigger = new CalTrigger().getNodeType(2) as CalTriggerV1;
 
 		mockHookFunctions = {
 			getNodeWebhookUrl: vi.fn(),
